requirements.txt
futures
pyserial
